streaming/app/requirements.txt
kafka-python
clickhouse-driver
python-dateutil
confluent-kafka